--- content/about-me/index-1.md
@@ -1,16 +1,0 @@
----
-about-page-title: About me
-description: fsadfsadfada
-social-media:
-  technologies:
-    - technology-name: dsada
-  tools:
-    - tool-name: dsada
-  other-skills:
-    - other-skill-name: dsada
-testimonials:
-  - testimonial-text: dsada
-    testimonial-image: image.png
-    testimonial-name: dsadasd
-    testimonial-job: dsadad
----
