+++ content/about/index-1.md
@@ -1,0 +1,59 @@
+---
+about-page-title: About
+description: Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce iaculis ultricies eros, ac iaculis eros maximus nec. Phasellus vitae mi felis. Curabitur iaculis nunc laoreet eleifend tincidunt. Aliquam nec lectus varius, mattis sapien nec, tristique elit. Suspendisse id sem sit amet mi tempus fermentum. Duis non metus nec nisl convallis efficitur. Donec condimentum neque rhoncus congue volutpat.  <br> Maecenas scelerisque varius ornare. Morbi egestas rutrum mattis. In pulvinar imperdiet tellus vel laoreet. Etiam ac sodales nulla. Nulla sit amet enim ipsum. Nulla pharetra turpis tempus venenatis sollicitudin. Nunc lacinia rhoncus leo, non consectetur velit tempus sed. Praesent euismod nibh ut massa congue, at condimentum justo egestas. Etiam accumsan felis non ipsum vulputate gravida. Etiam elementum vitae nisl a aliquam.
+social-media:
+  technologies:
+    - technology-name: HTML
+      technology-value: 5 
+    - technology-name: CSS
+      technology-value: 5 
+    - technology-name: JavaScript (ES9)
+      technology-value: 4
+    - technology-name: SASS
+      technology-value: 3
+    - technology-name: React.js
+      technology-value: 5
+    - technology-name: Redux
+      technology-value: 3
+    - technology-name: Storybook
+      technology-value: 2
+    - technology-name: React Testing Library
+      technology-value: 2
+    - technology-name: Gatsby
+      technology-value: 2
+    - technology-name: Apollo Client
+      technology-value: 3
+    - technology-name: Redux Saga
+      technology-value: 1
+    - technology-name: Next.js
+      technology-value: 2 
+  tools:
+    - tool-name: Git
+      tool-value: 2 
+    - tool-name: GitHub
+      tool-value: 3
+    - tool-name: JIRA
+      tool-value: 1
+    - tool-name: Trello
+      tool-value: 2
+    - tool-name: Figma
+      tool-value: 5
+    - tool-name: Zeplin
+      tool-value: 4
+  other-skills:
+    - other-skill-name: SCRUM
+      other-skill-value: 3
+    - other-skill-name: RWD
+      other-skill-value: 4
+    - other-skill-name: Atomic Design
+      other-skill-value: 2
+testimonials:
+  - testimonial-text: Mauris suscipit risus ut lorem consequat, id gravida sem vulputate. Donec pharetra mi ac elit hendrerit, at viverra risus fermentum. Proin sit amet viverra dolor.
+    testimonial-image: MarleneWatson.png
+    testimonial-name: Marlene Watson
+    testimonial-job: CEO of Donec
+  - testimonial-text: Suspendisse ullamcorper massa nec lectus volutpat malesuada. Proin at nibh dui. Nunc at tincidunt mauris, non sagittis libero. Proin porta id elit sit amet maximus.
+    testimonial-image: GabeMcdonalid.png
+    testimonial-name: Gabe Mcdonalid
+    testimonial-job: Marketing specialist in Railroad
+---
